# Conflicts:
#	src/main/java/com/ngng/api/global/security/handler/CustomLogoutHandler.java
#	src/main/java/com/ngng/api/global/security/jwt/filter/LoginFilter.java
#	src/main/java/com/ngng/api/global/security/jwt/filter/TokenFilter.java
#	src/main/java/com/ngng/api/user/service/JoinService.java

+++ src/main/java/com/ngng/api/global/security/handler/CustomLogoutHandler.java
@@ -2,12 +2,10 @@ package com.ngng.api.global.security.handler;
 
 import com.ngng.api.global.security.jwt.entity.Token;
 import com.ngng.api.global.security.jwt.repository.TokenRepository;
-import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Component;
 
@@ -19,10 +17,14 @@ public class CustomLogoutHandler implements LogoutHandler {
 
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-
-        // request에서 Authorization header(accessToken)을 받아와 해당 토큰이 db에 저장되어 있는지 검증
+        // TODO : Cookie로 수정
         String refreshToken = request.getHeader("Authorization");
-        Token token = tokenRepository.findTokenByTokenName(refreshToken).orElseThrow(() -> new JwtException("Invalid Token"));
+        Token token = tokenRepository.findTokenByTokenName(refreshToken);
+
+        if (token == null) {
+
+            return;
+        }
 
         tokenRepository.delete(token);
     }
